// src/types/purchase.ts
export interface PurchaseFormType {
  name: string;
  surname: string;
  phoneNumber: string;
  email: string;
  dateOfBirth: string;
  address: string;
  city: string;
  state: string;
  zipCode: string;
}

//Mocked purchase to the fake REST API service; Including autogenerated purchaseId, bought minifigId and filled shipping details
export interface Purchase {
  purchaseId: string;
  shippingDetails: PurchaseFormType;
  minifigId: string;
}